requirements.txt
bson
Django==1.8.3
freeze==1.0.9
requests==2.8.1
six==1.10.0
wheel==0.24.0
djangorestframework
mpmath
objectpath
python-usernames
drfdocs
status